requirements.txt
rich
pytest
flask
numpy
cryptography
argon2-cffi
Flask==3.0.3
Werkzeug==3.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
markupsafe==2.1.5
cryptography==43.0.1
